Manually ignore some paths in project.el
I'm not sure why this is the case, but due to the weird project structure in the
patch monorepo, I get a lot of junk.

Also, load the `magit` option upfront in the project.el dispatch, so I don't
have to find-file first, and invoke `magit-status` manually before it gets added
to the dispatch function.